@@ -1,73 +1,70 @@
 {
-  "name": "a2",
+  "name": "uctalks",
   "version": "0.0.0",
   "license": "MIT",
-  "angular-cli": {},
   "scripts": {
-    "start": "ng serve --hmr -e=dev-hmr",
-    "build-aot-and-deploy": "npm run build-aot && npm run deploy",
-    "test": "ng test",
-    "e2e": "ng e2e",
-    "lint": "ng lint",
     "ng": "ng",
-    "serve": "ng serve",
+    "start": "ng serve",
     "build": "ng build --prod",
     "build-aot": "ng build --prod --aot",
-    "deploy": "gh-pages -d dist -b master"
+    "build-aot-and-deploy": "npm run build-aot && npm run deploy",
+    "deploy": "gh-pages -d dist -b master",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.3.2",
-    "@angular/common": "4.3.2",
-    "@angular/compiler": "4.3.2",
-    "@angular/core": "4.3.2",
-    "@angular/forms": "4.3.2",
-    "@angular/http": "4.3.2",
-    "@angular/material": "2.0.0-beta.8",
-    "@angular/platform-browser": "4.3.2",
-    "@angular/platform-browser-dynamic": "4.3.2",
-    "@angular/router": "4.3.2",
+    "@angular/animations": "^5.2.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
+    "@angular/material": "^5.2.4",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/router": "^5.2.0",
+    "@angular/service-worker": "^5.2.0",
     "@ngrx/core": "1.2.0",
-    "@ngrx/db": "2.0.2",
-    "@ngrx/effects": "4.0.1",
-    "@ngrx/router-store": "4.0.0",
-    "@ngrx/store": "4.0.0",
-    "@ngrx/store-devtools": "4.0.0",
+    "@ngrx/db": "^2.1.0",
+    "@ngrx/effects": "^5.2.0",
+    "@ngrx/router-store": "^5.2.0",
+    "@ngrx/store": "^5.2.0",
+    "@ngrx/store-devtools": "^5.2.0",
     "angular2-jwt": "0.2.3",
     "auth0-js": "9.3.3",
-    "core-js": "2.4.1",
-    "global": "4.3.2",
+    "core-js": "^2.5.3",
     "hammerjs": "2.0.8",
-    "ng2-md-datatable": "1.4.0",
+    "ng2-md-datatable": "1.7.4",
+    "npm": "^5.7.1",
     "reselect": "3.0.1",
-    "rxjs": "5.4.2",
-    "ts-helpers": "1.1.2",
-    "zone.js": "0.8.16"
+    "rxjs": "^5.5.6",
+    "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cdk": "2.0.0-beta.8",
-    "@angular/cli": "1.2.6",
-    "@angular/compiler-cli": "4.3.2",
-    "@angular/language-service": "4.3.2",
-    "@angularclass/hmr": "2.1.3",
-    "@types/jasmine": "2.5.53",
-    "@types/node": "8.0.17",
-    "codelyzer": "3.1.2",
-    "common-tags": "1.4.0",
-    "eslint": "4.3.0",
-    "gh-pages": "1.0.0",
-    "jasmine-core": "2.7.0",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "ngrx-store-freeze": "0.1.9",
+    "@angular/cdk": "^5.2.4",
+    "@angular/cli": "~1.7.3",
+    "@angular/compiler-cli": "^5.2.0",
+    "@angular/language-service": "^5.2.0",
+    "@types/auth0-js": "^8.11.2",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/karma-jasmine": "^0.0.29",
+    "@types/node": "~6.0.60",
+    "codelyzer": "^4.0.1",
+    "gh-pages": "^1.1.0",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "^1.1.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ngrx-store-freeze": "^0.2.1",
     "protractor": "~5.1.2",
-    "ts-node": "3.3.0",
-    "tslint": "5.5.0",
-    "typescript": "2.4.2"
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "~2.5.3"
   }
 }
