@@ -1,3 +1,3 @@
 export const enum AutomationLocator {
   HeaderTitle = 'header.title',
-};
+}
