@@ -1,4 +1,4 @@
-import { browser, element, by } from 'protractor';
+import { browser} from 'protractor';
 import { AutomationLocator } from '../automation-locators.enum';
 import { getElementByAutomationLocator } from './utils';
 
