@@ -1,7 +1,6 @@
 // Protractor configuration file, see link for more information
 // https://github.com/angular/protractor/blob/master/lib/config.ts
 
-/*global jasmine */
 const { SpecReporter } = require('jasmine-spec-reporter');
 
 exports.config = {
@@ -9,11 +8,9 @@ exports.config = {
   specs: [
     './e2e/**/*.e2e-spec.ts'
   ],
-  multiCapabilities:[
-    {
-      'browserName': 'chrome'
-    }
-  ], 
+  capabilities: {
+    'browserName': 'chrome'
+  },
   directConnect: true,
   baseUrl: 'http://localhost:4200/',
   framework: 'jasmine',
@@ -22,12 +19,10 @@ exports.config = {
     defaultTimeoutInterval: 30000,
     print: function() {}
   },
-  beforeLaunch: function() {
-    require('ts-node').register({
-      project: 'e2e'
-    });
-  },
   onPrepare() {
+    require('ts-node').register({
+      project: 'e2e/tsconfig.e2e.json'
+    });
     jasmine.getEnv().addReporter(new SpecReporter({ spec: { displayStacktrace: true } }));
   }
 };
