@@ -1,30 +1,46 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
+import * as auth0 from 'auth0-js';
+import { AuthorizeOptions } from 'auth0-js';
 import 'rxjs/add/operator/filter';
 import 'rxjs/add/observable/bindNodeCallback';
 import { Observable } from 'rxjs/Observable';
-import auth0 from 'auth0-js';
-import { MdSnackBar } from '@angular/material';
 import { AUTH_CONFIG } from './auth-config';
-import { UserService } from '../user-service/user.service';
-import { Store } from '@ngrx/store';
-import { State } from 'app/reducers';
 
 @Injectable()
 export class AuthService {
   // Configure Auth0
   public auth0 = new auth0.WebAuth(AUTH_CONFIG);
 
-  constructor(
-    private router: Router,
-    private snackBar: MdSnackBar,
-    private store: Store<State>,
-    private userService: UserService,
-  ) { }
+  constructor(private router: Router,) {
+  }
+
+  public get isAuthenticated(): boolean {
+    // Check whether the current time is past the
+    // access token's expiry time
+    const expiresAt = JSON.parse(localStorage.getItem('expires_at'));
+    return new Date().getTime() < expiresAt;
+  }
+
+  public get userProfileId(): string | null {
+    return localStorage.getItem('userId');
+  }
+
+  public get userProfileName(): string | null {
+    return localStorage.getItem('userName');
+  }
+
+  public get userProfilePicture(): string | null {
+    return localStorage.getItem('userPicture');
+  }
 
   public login(): void {
-    // Call the show method to display the widget.
-    this.auth0.authorize({ connection: 'google-oauth2', prompt: 'select_account' });
+    /**
+     * Cast options to AuthorizeOptions since prompt property is needed for
+     * the google account selection but it is not present in AuthorizeOptions interface
+     */
+    const options = <AuthorizeOptions>{ connection: 'google-oauth2', prompt: 'select_account' };
+    this.auth0.authorize(options);
   }
 
   public handleAuthentication(): Observable<any> {
@@ -48,24 +64,5 @@ export class AuthService {
 
     // Go back to the home route
     this.router.navigate(['/home']);
-  }
-
-  public get isAuthenticated(): boolean {
-    // Check whether the current time is past the
-    // access token's expiry time
-    const expiresAt = JSON.parse(localStorage.getItem('expires_at'));
-    return new Date().getTime() < expiresAt;
-  }
-
-  public get userProfileId(): string | null {
-    return localStorage.getItem('userId');
-  }
-
-  public get userProfileName(): string | null {
-    return localStorage.getItem('userName');
-  }
-
-  public get userProfilePicture(): string | null {
-    return localStorage.getItem('userPicture');
   }
 }
